separated test from class module
